# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatGPT Auto Register API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus];

export const AccountStatus = {
  active: "active",
  failed: "failed",
  pending: "pending",
  unverified: "unverified",
} as const;

export interface Account {
  id: number;
  email: string;
  password: string;
  status: AccountStatus;
  proxyUsed?: string | null;
  createdAt: string;
  notes?: string | null;
}

export interface CreateAccountRequest {
  email: string;
  password: string;
  notes?: string;
}

export type ProxyProtocol = (typeof ProxyProtocol)[keyof typeof ProxyProtocol];

export const ProxyProtocol = {
  http: "http",
  https: "https",
  socks4: "socks4",
  socks5: "socks5",
} as const;

export type ProxyStatus = (typeof ProxyStatus)[keyof typeof ProxyStatus];

export const ProxyStatus = {
  active: "active",
  failed: "failed",
  untested: "untested",
} as const;

export interface Proxy {
  id: number;
  host: string;
  port: number;
  username?: string | null;
  password?: string | null;
  protocol: ProxyProtocol;
  status: ProxyStatus;
  lastChecked?: string | null;
}

export type CreateProxyRequestProtocol =
  (typeof CreateProxyRequestProtocol)[keyof typeof CreateProxyRequestProtocol];

export const CreateProxyRequestProtocol = {
  http: "http",
  https: "https",
  socks4: "socks4",
  socks5: "socks5",
} as const;

export interface CreateProxyRequest {
  host: string;
  port: number;
  username?: string;
  password?: string;
  protocol: CreateProxyRequestProtocol;
}

export type RegistrationTaskStatus =
  (typeof RegistrationTaskStatus)[keyof typeof RegistrationTaskStatus];

export const RegistrationTaskStatus = {
  idle: "idle",
  running: "running",
  paused: "paused",
  completed: "completed",
  failed: "failed",
} as const;

export interface RegistrationTask {
  id: number;
  name: string;
  status: RegistrationTaskStatus;
  totalAccounts: number;
  successCount: number;
  failedCount: number;
  useProxy: boolean;
  emailDomain?: string | null;
  createdAt: string;
  startedAt?: string | null;
  completedAt?: string | null;
  logs?: string[];
}

export interface CreateTaskRequest {
  name: string;
  totalAccounts: number;
  useProxy: boolean;
  emailDomain?: string;
}

export interface Stats {
  totalAccounts: number;
  activeAccounts: number;
  failedAccounts: number;
  totalProxies: number;
  activeProxies: number;
  runningTasks: number;
  successRate: number;
}

export type AddProxiesBatchBody = {
  proxies: string[];
};

export type AddProxiesBatch201 = {
  added: number;
  failed: number;
};